Fixes for file upload
