+++ universal-application-tool-0.0.1/app/assets/javascripts/azure_upload.ts
@@ -48,14 +48,10 @@ class AzureUploadController {
         if (err) {
           throw err
         }
-        console.log(resp)
-        this.setFileUploadMetadata(
-          redirectUrl,
-          azureUploadProps,
-          resp,
-          blockBlobUrl
+        this.setFileUploadMetadata(redirectUrl, azureUploadProps, resp)
+        window.addEventListener('load', () =>
+          window.location.replace(redirectUrl.toString())
         )
-        window.location.replace(redirectUrl.toString())
       })
   }
 
@@ -77,14 +73,12 @@ class AzureUploadController {
   private setFileUploadMetadata(
     redirectUrl: URL,
     azureUploadProps: any,
-    resp: any,
-    blockBlobUrl: any
+    resp: any
   ) {
     redirectUrl.searchParams.set('originalFileName', azureUploadProps.file.name)
     redirectUrl.searchParams.set('etag', resp.eTag)
     redirectUrl.searchParams.set('fileName', azureUploadProps.fileName)
     redirectUrl.searchParams.set('container', azureUploadProps.containerName)
-    redirectUrl.searchParams.set('blockBlobUrl', blockBlobUrl)
   }
 }
